% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.hash.R
\name{add_hash}
\alias{add_hash}
\title{Merge cell hashcoding demultiplexed data with GEX Seurat 3 object with VDJ metadata}
\usage{
add_hash(hash.matrix, Seurat_obj, sample_name, quantile_theshold = 0.99)
}
\arguments{
\item{hash.matrix}{a cellular bacrode vs hashtags matrix}

\item{Seurat_obj}{Seurat object}
}
\value{
Seurat 3 object merged with hashtag demiltiplexed identities in the metadata slot
}
\description{
This function allows you to fuse the output of hashcode demultiplexing with GEX+VDJ object
}
\examples{

A07 <- add_hash(Seurat_objhash, Seurat_obj, 'A14')


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
