% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vdj.R
\name{add_vdj}
\alias{add_vdj}
\title{Merge clonotype and isotype data to GEX Seurat object}
\usage{
add_vdj(vdj_location, Seurat_obj)
}
\arguments{
\item{vdj_location}{location of cel ranger vdj output}

\item{Seurat_obj}{Seurat object}
}
\value{
umap 2d plot with velocity
}
\description{
This function allows you to fuse the output of VDJ library to Seurat object
}
\examples{

A07 <- add_vdj('path_to_vdj', A07)


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
