% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_markers.R
\name{annotate_markers}
\alias{annotate_markers}
\title{Annotation for the cell markers.}
\usage{
annotate_markers(markers_table, X = 50, organism = "hsapiens")
}
\arguments{
\item{markers_table}{The output of the Seurat::FndAllMarkers function, don't change anything in the column names within it.}

\item{X}{How much of the genes you need to analyse in each cluster. 50 is a default.}

\item{organism}{What species you work with? "mmusculus" and "hsapiens" is available in gProfiler. Human is a default.}
}
\description{
This function allows you to get multiple functional annotations for DE genes for each cell cluster using a Seurat::FindAllMarkers function output
}
\examples{
annotate_markers(stromal_markers)
}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
