% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.cell.corheatmap.R
\name{Cell.cell.corheatmap}
\alias{Cell.cell.corheatmap}
\title{heatmap for cell-cell correlation matrix}
\usage{
Cell.cell.corheatmap(Seurat_obj, do.cluster = F, colours = "random")
}
\arguments{
\item{Seurat_obj}{Seurat object}

\item{do.cluster}{set to T if you want a hierarchical clustering on the top of the heatmap, defulat is FALSE}

\item{colours}{set the colours for clusters that you want, random colours by default}
}
\value{
a heaatmap
}
\description{
This function allows you to plot a heatmap for cell-cell correlation
}
\examples{

Cell.cell.corheatmap(subSeurat_obj, do.cluster = T)

}
\keyword{analysis,}
\keyword{cell,}
\keyword{correlation}
\keyword{data}
\keyword{heatmap,}
\keyword{single}
\keyword{visualisation}
