% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge and clean loom.R
\name{clean_spmat}
\alias{clean_spmat}
\title{Run Velocyto analysis on your Seurat2 object}
\usage{
clean_spmat(mat, Seurat_obj)
}
\arguments{
\item{mat}{matrix}

\item{emb}{Seurat_obj Seurat object}
}
\value{
cleaned matrix
}
\description{
This function allows you to get 2d embeddings from your Seurat object that you want to use for velocyto analysis
}
\examples{



}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
