% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.R
\name{clonotypes_summary}
\alias{clonotypes_summary}
\title{Gives you a summary of your clonotypes}
\usage{
clonotypes_summary(Seurat_obj, clusters_column)
}
\arguments{
\item{clusters_column}{column in the @meta.data that stores your clusters}

\item{Seuraat_obj}{Seurat object}
}
\value{
data.frame with clonotypes summary
}
\description{
Gives you a summary of your clonotypes
}
