% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_and_put.signature.R
\name{complete}
\alias{complete}
\title{gets you an overlap of gene list with your object}
\usage{
complete(markers, Seurat_obj)
}
\arguments{
\item{markers}{yout gene set, it can be a set of markers for cell type identification or genes involved in the certain pathway that you want localise}

\item{Seurat_obj}{your Seurat object}
}
\value{
overlapping gene list
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\examples{

Ccl19hi.TRC <- complete(rodda$Ccl19hi.TRC, PP.combined)

}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
