% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dot_plot_topXgenes}
\alias{dot_plot_topXgenes}
\title{Make a dot plot for all your markers}
\usage{
dot_plot_topXgenes(markers_table, X = 5, Seurat_obj)
}
\arguments{
\item{markers_table}{The output of the Seurat::FndAllMarkers function, don't change anything in the column names within it.}

\item{X}{How much of the genes you need to analyse in each cluster. 5 is a default.}

\item{Seurat_obj}{your Seurat object to plot markers from}
}
\description{
This function allows you to plot each top X DEG for each cell cluster using a Seurat::FindAllMarkers function output
}
\examples{

dot_plot_topXgenes(markers, 10)

}
\keyword{RNA-seq}
\keyword{Seurat,}
\keyword{cell}
\keyword{dot}
\keyword{markers,}
\keyword{plot,}
\keyword{sequencing,}
\keyword{single}
