% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene.gene.corheatmap.R
\name{Gene.gene.corheatmap}
\alias{Gene.gene.corheatmap}
\title{heatmap for gene-gene correlation matrix}
\usage{
Gene.gene.corheatmap(Seurat_obj, cor.method = "spearman",
  cor.cut.off = 0.3)
}
\arguments{
\item{Seurat_obj}{Seurat object}
}
\value{
a heaatmap
}
\description{
This function allows you to plot a heatmap for gene-gene correlation
}
\examples{

Gene.gene.corheatmap(Seurat_obj, do.cluster = T)

}
\keyword{analysis,}
\keyword{cell,}
\keyword{correlation}
\keyword{data}
\keyword{heatmap,}
\keyword{single}
\keyword{visualisation}
