% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_contribution.R
\name{Gene_contribution}
\alias{Gene_contribution}
\title{Calculate the contribution of each gene for the given gene signature}
\usage{
Gene_contribution(Seurat_obj, ident, signature, signature_name)
}
\arguments{
\item{Seurat_obj}{your Seurat object}

\item{ident}{number of cluster you want to look at}

\item{signature}{the gene set of your interest}
}
\value{
heatmap with the gene contribution
}
\description{
This function allows you to see how did each gene from the given transcriptional program contribute to the pre-calculated gene signature
}
\examples{

Gene_contribution(A07, 7, blooddata[[1]])

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{gene}
\keyword{sequencing,}
\keyword{signature}
\keyword{single}
