% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{ggplot extension for split violin plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\value{
ggplot extension
}
\description{
ggplot extension for split violin plot
}
\examples{

ggplot(data, aes(x=x, y=y, fill = factor))+
geom_split_violin

}
\keyword{ggplot}
\keyword{plot,}
\keyword{violin}
