% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_velocyto.R
\name{get_emb}
\alias{get_emb}
\title{Run Velocyto analysis on your Seurat2 object}
\usage{
get_emb(Seurat_obj, emb = c("umap", "monocle", "custom"))
}
\arguments{
\item{Seurat_obj}{Seurat object}

\item{emb}{embedding type of your choice}
}
\value{
umap 2d plot with velocity
}
\description{
This function allows you to get 2d embeddings from your Seurat object that you want to use for velocyto analysis
}
\examples{

emb <- get_emb(B06, emb = 'umap')

rvel.cd <- preprocess_loom('~/Desktop/RIMMI/Adam_germinal_cenrte/B06/velocyto/B06.loom', B06, emb)

Seurat2_velocyto(rvel.cd, B06, emb)


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
