% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_DEG.RData.R
\name{Hglm.deg}
\alias{Hglm.deg}
\title{apply DEG with multiple glms on Seruat object with case control meta data}
\usage{
Hglm.deg(
  Seurat_obj,
  condition = "condition",
  cell.types = "RNA_snn_res.0.5",
  model = c("glm0", "glm", "glmer")
)
}
\arguments{
\item{Seurat_obj}{your Seurat object}

\item{model}{type of the model to choose}
}
\value{
DEG table
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
