% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.hash.R
\name{HTODemux}
\alias{HTODemux}
\title{HTODemux
in house modified Seurat::HTODemux for identifiyng multuplets}
\usage{
HTODemux(
  object,
  assay = "HTO",
  positive.quantile = 0.99,
  init = NULL,
  nstarts = 100,
  kfunc = "clara",
  nsamples = 100,
  seed = 42,
  verbose = TRUE,
  patient = "patient1"
)
}
\arguments{
\item{object}{Seurat object}
}
\description{
HTODemux
in house modified Seurat::HTODemux for identifiyng multuplets
}
