% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotypes_DEG.R
\name{Isotypes_DEG}
\alias{Isotypes_DEG}
\title{This function allows you to calculate DEG for the IG Isotypes}
\usage{
Isotypes_DEG(Seurat_obj)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}
}
\description{
This function allows you to calculate DEG for the IG Isotypes
}
\examples{

Isotypes_DEG(A07)

}
\keyword{B}
\keyword{DEG,}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cell,}
\keyword{cells}
\keyword{cells,}
\keyword{single}
