% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lib.qc_plot}
\alias{lib.qc_plot}
\title{Plot clusters in 2 umaps with the point size corresponting to the library size}
\usage{
lib.qc_plot(Seurat_obj)
}
\arguments{
\item{Seurat_obj}{your Seurat object}
}
\value{
scatter plot with the library size correponding to the point size
}
\description{
This function allows you to see if the library size influences in your clustering
}
\examples{

scatter_libdepth(A07)

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{gene}
\keyword{sequencing,}
\keyword{signature}
\keyword{single}
