% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_DEG.RData.R
\name{list.to.table}
\alias{list.to.table}
\title{unite list of deg table}
\usage{
list.to.table(deg.table)
}
\arguments{
\item{deg.table}{deg table}
}
\value{
deg table united
}
\description{
better version volcano plot
}
\keyword{DEG,}
\keyword{analysis}
\keyword{differential}
\keyword{expression}
