% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_to_xls.R
\name{markers_to_xls}
\alias{markers_to_xls}
\title{getting your markers into the excel shpreadsheet}
\usage{
markers_to_xls(markers_table, filename = "my_markers")
}
\arguments{
\item{markers_table}{The output of the Seurat::FndAllMarkers function, don't change anything in the column names within it.}
}
\description{
This function allows you write Seurat::FindAllMarkers function output into the excel file having each cluster markers written in the distinct sheet
}
\examples{
annotate_markers(my_markers, "stromal_markers")
}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell,}
\keyword{excel,}
\keyword{markers}
\keyword{single}
