% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge and clean loom.R
\name{merge_loom}
\alias{merge_loom}
\title{Merge two loom files into one and produce rvel.cd}
\usage{
merge_loom(loom1, loom2, Seurat_obj, sample1, sample2, emb)
}
\arguments{
\item{loom1}{path to loom file 1}

\item{loom2}{path to loom file 2}

\item{sample1}{sample name 1, for the prefix in cellular barcodes}

\item{sample2}{sample name 2, for the prefix in cellular barcodes}

\item{emb}{2d embedding}
}
\value{
umap 2d plot with velocity
}
\description{
This function allows you to merge emat and nmat matrices into one and then calculate rvel.cd for velocyto analysis
}
\examples{



}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
