% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_velocyto.R
\name{preprocess_loom}
\alias{preprocess_loom}
\title{Run Velocyto analysis on your Seurat2 object}
\usage{
preprocess_loom(loom_path, Seurat_obj, emb)
}
\arguments{
\item{loom_path}{path to the loom file, pre-calculated with python command line tool}

\item{Seurat_obj}{Seurat object}

\item{emb}{the matrix with 2d embeddings that you got from get_emb() function}
}
\value{
umap 2d plot with velocity
}
\description{
This function allows you to preprocess loom file into the object that fits to plotting velocyto arrows on 2d embeddings
}
\examples{

emb <- get_emb(B06, emb = 'umap')

rvel.cd <- preprocess_loom('~/Desktop/RIMMI/Adam_germinal_cenrte/B06/velocyto/B06.loom', B06, emb)

Seurat2_velocyto(rvel.cd, B06, emb)


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
