% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_DEG.RData.R
\name{pseudo.bulk.DEG}
\alias{pseudo.bulk.DEG}
\title{apply DESeq for DEG pseudo bulk collapsed samples}
\usage{
pseudo.bulk.DEG(Seurat_sub.obj, thresh.pv = 0.4,
  show.genes = c("CCL19", "CCL21A"))
}
\arguments{
\item{thresh.pv}{display genes with at least this p value}

\item{show.genes}{additional genes to display}

\item{Seurat_obj}{your Seurat object}
}
\value{
DEG table
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
