% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotypes_DEG.R
\name{pull_clonotype}
\alias{pull_clonotype}
\title{This function allows you to pull specific barcodes from the object with specific clonotypes from specific clusters}
\usage{
pull_clonotype(Seurat_obj, clonotypes, clusters = NULL)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}

\item{clonotypes}{clonotype names}

\item{clusters}{cluster names}
}
\description{
This function allows you to pull specific barcodes from the object with specific clonotypes from specific clusters
}
\examples{

pull_clonotype(A07, 'IGHA1')

}
\keyword{B}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cells}
\keyword{cells,}
