% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashtags_DEG.R
\name{pull_hashtag}
\alias{pull_hashtag}
\title{This function allows you to pull specific barcodes from the object with specific IG hashtags from specific clusters}
\usage{
pull_hashtag(Seurat_obj, hashtags, demult.method = "HTODemux", clusters = NULL)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Ihashtag assigment}

\item{hashtags}{Ihashtag names}

\item{clusters}{cluster names}
}
\description{
This function allows you to pull specific barcodes from the object with specific IG hashtags from specific clusters
}
\examples{

pull_ihashtag(A07, 'IGHA1')

}
\keyword{B}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cells}
\keyword{cells,}
