% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotypes_DEG.R
\name{pull_isotype}
\alias{pull_isotype}
\title{This function allows you to pull specific barcodes from the object with specific IG Isotypes from specific clusters}
\usage{
pull_isotype(Seurat_obj, isotypes, clusters = NULL)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}

\item{isotypes}{Isotype names}

\item{clusters}{cluster names}
}
\description{
This function allows you to pull specific barcodes from the object with specific IG Isotypes from specific clusters
}
\examples{

pull_isotype(A07, 'IGHA1')

}
\keyword{B}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cells}
\keyword{cells,}
