% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vdj.R
\name{pull_vdj}
\alias{pull_vdj}
\title{Merge clonotype and isotype data to GEX Seurat object}
\usage{
pull_vdj(vdj_location)
}
\arguments{
\item{vdj_location}{location of cel ranger vdj output}
}
\value{
table with vdj info
}
\description{
This function allows you to fuse the output of VDJ library to Seurat object
}
\examples{

A07 <- add_vdj('path_to_vdj', A07)


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
