% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_and_put.signature.R
\name{put_signature}
\alias{put_signature}
\title{Calculate average expression for gene set.}
\usage{
put_signature(markers, Seurat_obj, title)
}
\arguments{
\item{markers}{yout gene set, it can be a set of markers for cell type identification or genes involved in the certain pathway that you want localise}

\item{Seurat_obj}{your Seurat object}

\item{title}{the name of your gene set}
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\examples{
put_signature(markers = c("CCL21", "CCL19", "TNFSF13B"), Seurat_obj = human_stromal, title = "T-zone markers")
}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
