% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_genesets.R
\name{quantify_genesets}
\alias{quantify_genesets}
\title{This function allows you to calculate an average expression for the list of genes of your interest.}
\usage{
quantify_genesets(Seurat_obj, genesets, reset_genesets = FALSE)
}
\arguments{
\item{Seurat_obj}{Seurat: your Seurat object}

\item{genesets, }{list: named list with genesets}

\item{reset_genesets, }{bool: if you want reset already existing genesets with a new assay object}
}
\value{
Seurat object with geneset assay
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\examples{

}
\keyword{gene}
\keyword{geneset,}
\keyword{signature,}
