% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query_heatmap.R
\name{query_heatmap}
\alias{query_heatmap}
\title{heatmap genes with certain function}
\usage{
query_heatmap(query, Seurat_obj, species = "Mouse")
}
\arguments{
\item{query}{string containing a regular expression to find a certain gene function}

\item{Seurat_obj}{Seurat object}
}
\value{
heatmap
}
\description{
This function allows you to plot a heatmap of some genes with the certain function
}
\examples{

query_heatmap('chemokine', pp.comp)

}
\keyword{function}
\keyword{gene}
\keyword{genes,}
\keyword{heatmap,}
