% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.R
\name{rank_clonotypes}
\alias{rank_clonotypes}
\title{Rank clonotypes based of the spread in GEX space}
\usage{
rank_clonotypes(Seurat_obj, cells_th = 10)
}
\arguments{
\item{cells_th}{minimum cells in the clonotype to analyse}

\item{Seuraat_obj}{Seurat object}
}
\value{
data.frame ranked clonotypes
}
\description{
Rank clonotypes based of the spread in GEX space
}
