% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save.plot_grid}
\alias{save.plot_grid}
\title{save multiple gg plots in pdf}
\usage{
save.plot_grid(..., filename)
}
\arguments{
\item{...}{ggplot objects}

\item{filename}{filename}
}
\description{
This function allows you to save multiple gg plots in pdf
}
\examples{
p1 <- qplot(1:10)
p2 <- qplot(1:20)
p3 <- qplot(1:30)

save.plot.grid(p1,p2,p3, 'whatever')


}
\keyword{ggplot,}
\keyword{save}
