% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_PAGA.R
\name{Seurat2_PAGA}
\alias{Seurat2_PAGA}
\title{Run PAGA analysis on your Seurat2 object}
\usage{
Seurat2_PAGA(Seurat_obj)
}
\arguments{
\item{Seurat_obj}{Seurat object}
}
\value{
umap plot, Seurat object with PAGA information
}
\description{
This function allows you to run PAGA analysis on your Seurat2 object, visualise and then store in your Seurat object
}
\examples{


}
\keyword{PAGA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{analysis,}
\keyword{cell}
\keyword{pseudotime}
\keyword{sequencing,}
\keyword{single}
