% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_velocyto.R
\name{Seurat2_velocyto}
\alias{Seurat2_velocyto}
\title{Run Velocyto analysis on your Seurat2 object}
\usage{
Seurat2_velocyto(loom_path, Seurat_obj)
}
\arguments{
\item{loom_path}{path to the loom file, pre-calculated with python command line tool}

\item{Seurat_obj}{Seurat object}
}
\value{
umap 2d plot with velocity
}
\description{
This function allows you to un Velocyto analysis on your Seurat2 object and visualise it on your umap embeddings
}
\examples{



}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
