% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_PAGA.R
\name{Seurat_obj2_PAGA}
\alias{Seurat_obj2_PAGA}
\title{Run PAGA analysis on your Seurat_obj2 object}
\usage{
Seurat_obj2_PAGA(Seurat_obj_obj,
  python_path = "/usr/local/bin/anaconda3/bin/python3.6")
}
\arguments{
\item{Seurat_obj_obj}{Seurat_obj object}
}
\value{
umap plot, Seurat_obj object with PAGA information
}
\description{
This function allows you to run PAGA analysis on your Seurat_obj2 object, visualise and then store in your Seurat_obj object
}
\examples{


}
\keyword{PAGA}
\keyword{RNA-seq,}
\keyword{Seurat_obj,}
\keyword{analysis,}
\keyword{cell}
\keyword{pseudotime}
\keyword{sequencing,}
\keyword{single}
