% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hashtags_DEG.R
\name{Specific_hashtags_DEG}
\alias{Specific_hashtags_DEG}
\title{This function allows you to calculate DEG for some specific IG hashtags}
\usage{
Specific_hashtags_DEG(Seurat_obj, hash1, hash2, filename,
  demult.method = "HTODemux", clusters = NULL)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Ihashtag assigment}
}
\description{
This function allows you to calculate DEG for some specific IG hashtags
}
\examples{

hashtags_DEG(A07)

}
\keyword{B}
\keyword{DEG,}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cell,}
\keyword{cells}
\keyword{cells,}
\keyword{single}
