% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotypes_DEG.R
\name{Specific_isotypes_DEG}
\alias{Specific_isotypes_DEG}
\title{This function allows you to calculate DEG for some specific IG Isotypes}
\usage{
Specific_isotypes_DEG(Seurat_obj, iso1, iso2, filename, clusters = NULL)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}
}
\description{
This function allows you to calculate DEG for some specific IG Isotypes
}
\examples{

Isotypes_DEG(A07)

}
\keyword{B}
\keyword{DEG,}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cell,}
\keyword{cells}
\keyword{cells,}
\keyword{single}
