% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.hash.R
\name{throw_away_tags}
\alias{throw_away_tags}
\title{mark tags that you think are useless as such}
\usage{
throw_away_tags(Seurat_obj, useless_tags)
}
\arguments{
\item{Seurat_obj}{Seurat object}

\item{useless_tags}{indexes of tags that you think are useless}
}
\value{
Seurat 3 object
}
\description{
This function allows you to  mark tags that you think are useless as such
}
\examples{

throw_away_tags(Seurat_obj, 
useless_tags =  c(1,2,3))


}
\keyword{RNA}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{velocity}
