% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot3d.R
\name{umap_plot3d}
\alias{umap_plot3d}
\title{Plot your clusters in 3 dimensions}
\usage{
umap_plot3d(obj, dims = 1:10)
}
\arguments{
\item{obj}{your Seurat object}

\item{dims}{how many of pca/cca dimentions you want to use, 10 is default}
}
\value{
plot
}
\description{
This function allows you to make a 3D scatter plot for your single cell gene expression measurements using UMAP dimensions by default from your Seurat object
}
\examples{


umap_plot3d(tcells)

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{dimentionality}
\keyword{reduction}
\keyword{sequencing,}
\keyword{single}
\keyword{uMAP,}
