% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uMAPPlot.R
\name{uMAPPlot}
\alias{uMAPPlot}
\title{Plot uMAP in 2 dimentions}
\usage{
uMAPPlot(Seurat_obj, sample_name = "X")
}
\arguments{
\item{Seurat_obj}{your Seurat object}

\item{title}{title for your plot}
}
\description{
This function allows you in addition to plot your clustering results in 2 uMAPs dimentions
}
\examples{
Seurat_obj <- Run_uMAP(Seurat_obj)
uMAPPlot(Seurat_obj, sample_name = 'b06')
}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{dimentionality}
\keyword{reduction}
\keyword{sequencing,}
\keyword{single}
\keyword{uMAP,}
