% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unarchive_10x}
\alias{unarchive_10x}
\title{Prepare 10x v3 out files files for Seurat 2}
\usage{
unarchive_10x(path)
}
\arguments{
\item{path}{path to filtered gene bc matrix}
}
\description{
Prepare 10x v3 out files files for Seurat 2
}
\keyword{10x}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{sequencing,}
\keyword{single}
\keyword{unarchive('"~/.../filtered_feature_bc_matrix/"')}
