% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_DEG.RData.R
\name{volcano_pl2}
\alias{volcano_pl2}
\title{plot volcano plot 2}
\usage{
volcano_pl2(response, ES_th = 0.5, title)
}
\arguments{
\item{response}{table with DEG}

\item{ES_th}{ES}

\item{title}{the name of your plot}
}
\value{
volcano plot
}
\description{
better version volcano plot
}
\examples{

volcano_pl(responce)

}
\keyword{DEG,}
\keyword{analysis,}
\keyword{differential}
\keyword{expression}
\keyword{plot}
\keyword{volcano}
