% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{plor volcano plot}
\usage{
volcano_plot(edgeR_output, title = "volcano plot")
}
\arguments{
\item{edgeR_output}{your edgeR output}

\item{title}{the name of your plot}
}
\value{
volcano plot
}
\description{
This function allows you to vizualise differential expression analisys for bulk stidy by volcano plot
}
\examples{

volcano_plot(edgeR_output)

}
\keyword{DEG,}
\keyword{analysis,}
\keyword{differential}
\keyword{expression}
\keyword{plot}
\keyword{volcano}
