% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vdj.R
\name{clonotypes_summary}
\alias{clonotypes_summary}
\title{This function allows you pull a summary information about your clonotypes}
\usage{
clonotypes_summary(Seurat_obj, n = 10, separate_cdr3 = T)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}

\item{n}{how much clones}
}
\description{
This function allows you pull a summary information about your clonotypes
}
\examples{

clonotypes_summary(A07, n = 10)

}
\keyword{B}
\keyword{DEG,}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cell,}
\keyword{cells}
\keyword{cells,}
\keyword{single}
