% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonotype_analysis.R, R/add_vdj.R
\name{clonotypes_summary}
\alias{clonotypes_summary}
\title{Gives you a summary of your clonotypes}
\usage{
clonotypes_summary(Seurat_obj, n = 10, separate_cdr3 = T)

clonotypes_summary(Seurat_obj, n = 10, separate_cdr3 = T)
}
\arguments{
\item{Seurat_obj}{Seurat object with integrated data from VDJ analysis with Isotype assigment}

\item{n}{how much clones}

\item{Seuraat_obj}{Seurat object}

\item{clusters_column}{column in the @meta.data that stores your clusters}
}
\value{
data.frame with clonotypes summary
}
\description{
Gives you a summary of your clonotypes

This function allows you pull a summary information about your clonotypes
}
\examples{

clonotypes_summary(A07, n = 10)

}
\keyword{B}
\keyword{DEG,}
\keyword{T}
\keyword{VDJ}
\keyword{analysis,}
\keyword{cell,}
\keyword{cells}
\keyword{cells,}
\keyword{single}
