% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene.gene.corheatmap.R
\name{Gene.gene.corheatmap}
\alias{Gene.gene.corheatmap}
\title{heatmap for gene-gene correlation matrix}
\usage{
Gene.gene.corheatmap(Seurat_obj, cor.method = "spearman",
  coef.cut.off = 0.3, gene.cut.off = 1, imputed = F,
  gene.names = T)
}
\arguments{
\item{Seurat_obj}{Seurat object}

\item{cor.method}{method to par ro cor function for correlation calculation, spearman is by default}

\item{coef.cut.off}{what monimum correlation coeffitient to choose to cut off the noise}

\item{gene.cut.off}{how much genes should have this correlation coefficient}

\item{imputed}{should MAGIC imputetion be used on an expression matrix primarily to correlation analysis, FALSE by default}

\item{gene.names}{do you want do show gene names? Set to FALSE is you gonna have a large matrix}
}
\value{
a heaatmap
}
\description{
This function allows you to plot a heatmap for gene-gene correlation
}
\examples{

Gene.gene.corheatmap(Seurat_obj, do.cluster = T)

}
\keyword{analysis,}
\keyword{cell,}
\keyword{correlation}
\keyword{data}
\keyword{heatmap,}
\keyword{single}
\keyword{visualisation}
