% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_contribution.R
\name{Gene_contribution}
\alias{Gene_contribution}
\title{Calculate the contribution of each gene for the given gene signature}
\usage{
Gene_contribution(Seurat_obj, resolution, signature, signature_name)
}
\arguments{
\item{Seurat_obj}{your Seurat object}

\item{resolution}{what resolution you wanna check, either 0.6 or 1 or 2}

\item{signature}{the gene set of your interest}

\item{signature_name}{the name you gave to the signature}
}
\value{
heatmap with the gene contribution
}
\description{
This function allows you to see how did each gene from the given transcriptional program contribute to the pre-calculated gene signature
}
\examples{

Gene_contribution(A07, 0.6, signature, 'sLZ')

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{gene}
\keyword{sequencing,}
\keyword{signature}
\keyword{single}
