% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_and_put.signature.R
\name{put_signature}
\alias{put_signature}
\title{Calculate average expression for gene set.}
\usage{
put_signature(markers, Seurat_obj, title)
}
\arguments{
\item{markers}{yout gene set, it can be a set of markers for cell type identification or genes involved in the certain pathway that you want localise}

\item{Seurat_obj}{your Seurat object}

\item{title}{the name of your gene set}
}
\value{
Seurat object with new feature, your gene set average expression
}
\description{
This function allows you to calculate an average expression for the list of genes of your interest.
}
\examples{

put_signature(markers = c("CCL21", "CCL19", "TNFSF13B"), Seurat_obj = human_stromal, title = "T-zone markers")
to plot it use FeaturePlot function with min and ,ax cutoff parameters being q3 and q97, it helps you to vizualise it without noise
FeaturePlot(Seurat_obj,
"T-zone markers",
cols.use = c("grey", "blue"),
min.cutoff = "q3",
max.cutoff = "q97")

}
\keyword{GO,}
\keyword{HPA,}
\keyword{KEGG,}
\keyword{RNA-seq}
\keyword{cell,}
\keyword{single}
