% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_uMAP.R
\name{Run_uMAP}
\alias{Run_uMAP}
\title{Calculate uMAP for Seurat object}
\usage{
Run_uMAP(Seurat_obj, variable_genes = TRUE)
}
\arguments{
\item{Seurat_obj}{your Seurat object}
}
\description{
This function allows you in addition to classical dimentionality reductions approaches as tSNE and PCA calculate also uMAP and embedd it to your Seurat object in the dr slot.
}
\examples{
Seurat_obj <- Run_uMAP(Seurat_obj)
uMAPPlot(Seurat_obj, sample_name = 'b06')
}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{dimentionality}
\keyword{reduction}
\keyword{sequencing,}
\keyword{single}
\keyword{uMAP,}
