% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_uMAP.R
\name{Run_uMAP}
\alias{Run_uMAP}
\title{Calculate uMAP for Seurat object}
\usage{
Run_uMAP(Seurat_obj, dims = 10, method = "naive", CCA = F)
}
\arguments{
\item{Seurat_obj}{your Seurat object}

\item{dims}{how many of pca dimentions you want to use, 10 is default}

\item{method}{method for umap manifold calculation, can be, "naive" or "umap-learn". Naive method is default. Further reading is in ??umap::umap}

\item{CCA}{set TRUE of you want to build uMAPs on the CCAs dimentionality rediction}
}
\value{
Seurat object with new dimentionality reduction slot
}
\description{
This function allows you in addition to canonical dimentionality reductions approache as tSNE calculate also uMAP and embedd it to your Seurat object in the dr slot, can be used for visualisation of the cell clusters.
}
\examples{

Seurat_obj <- Run_uMAP(Seurat_obj)
uMAPPlot(Seurat_obj, sample_name = 'b06')

}
\keyword{RNA-seq,}
\keyword{Seurat,}
\keyword{cell}
\keyword{dimentionality}
\keyword{reduction}
\keyword{sequencing,}
\keyword{single}
\keyword{uMAP,}
