% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_pl}
\alias{volcano_pl}
\title{plot volcano plot 2}
\usage{
volcano_pl(response, logFC_th = 0.5, title)
}
\arguments{
\item{response}{table with DEG}

\item{logFC_th}{LogFC threshold}

\item{title}{the name of your plot}
}
\value{
volcano plot
}
\description{
better version volcano plot
}
\examples{

volcano_pl(responce)

}
\keyword{DEG,}
\keyword{analysis,}
\keyword{differential}
\keyword{expression}
\keyword{plot}
\keyword{volcano}
